New Donor Annual Report
I have added a new class/controller that will run a report and show
each donor, with the total amount of donations for the year and
information about the donor. This report prints on a pdf or it can be
downloaded to a css file. There are two visual force pages for this
report using the same controller. Also the test class for the
controller, and a couple of changes on the visualforce page for the QA
reports and controller to include the button for the Annual report and
the methods to call them.