QA Quartery Volunteers Report
I changed the QAmonthlyVol class to be able to do the quarterly report
for volunteers as well. For this also the QAMonthlyVF visualforce page
was changed as well as the QAReportsController (so that it will call
the correct method). I added to the URL call a ‘q’ variable if q=N then
the report is a Monthly report.